back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["bastienlc <bastien.le-chenadec@eleves.enpc.fr>"]

[tool.poetry.dependencies]
python = ">3.9.0,<3.9.1 || >3.9.1,<4"
Django = "3.2.22"
Pillow = "10.0.1"
django-tellme = "0.7.3"
six = "1.16.0"
django-sendgrid-v5 = "1.2.2"
django-environ = "0.11.2"
gunicorn = "21.2.0"
django-cas-ng = "5.0.0"
django-markdownify = "^0.9.0"
djangorestframework = "3.14.0"
Unidecode = "^1.3.2"
django-storages = "^1.12.3"
django-cors-headers = "4.2.0"
django-reverse-js = "^0.1.3"
rest-pandas = "^1.1.0"
beautifulsoup4 = "^4.11.1"
celery = "^5.2.7"
redis = "5.0.1"
django-celery-beat = "2.5.0"
requests = "^2.26.0"
urllib3 = "2.0.6"
exponent_server_sdk = "2.0.0"
psycopg2-binary = "^2.9.10"
cryptography = "^46.0.1"
[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.2"
isort = "^6.0.1"

